pyproject.toml
[project]
name = "delete-binaries-deleted"
version = "0.1.0"
description = "Tool to safely delete all records from Binaries_deleted table in batches"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "rich>=14.0.0",
]
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

[project.scripts]
delete-binaries = "delete_binaries_deleted.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
